RB-497 New module: Project/Kanban

+++ src/main/java/com/rebuild/server/business/rbstore/RBStore.java
@@ -7,11 +7,9 @@ See LICENSE and COMMERCIAL in the project root for license information.
 
 package com.rebuild.server.business.rbstore;
 
-import cn.devezhao.commons.CodecUtils;
 import com.alibaba.fastjson.JSON;
 import com.rebuild.server.RebuildException;
-import com.rebuild.server.helper.SysConfiguration;
-import com.rebuild.utils.CommonsUtils;
+import com.rebuild.utils.HttpUtils;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
@@ -64,14 +62,16 @@ public class RBStore {
 		if (!fileUrl.startsWith("http")) {
 			fileUrl = DATA_REPO + fileUrl;
 		}
-		
-		File tmp = SysConfiguration.getFileOfTemp("rbstore." + CodecUtils.randomCode(6));
+
 		JSON d = null;
 		try {
-			CommonsUtils.readBinary(fileUrl, tmp);
-			String t2str = FileUtils.readFileToString(tmp, "utf-8");
-			d = (JSON) JSON.parse(t2str);
-			FileUtils.deleteQuietly(tmp);
+			File tmp = HttpUtils.readBinary(fileUrl);
+			if (tmp != null) {
+				String t2str = FileUtils.readFileToString(tmp, "utf-8");
+				d = (JSON) JSON.parse(t2str);
+				FileUtils.deleteQuietly(tmp);
+			}
+
 		} catch (Exception e) {
 			LOG.error("Fetch failure from URL : " + fileUrl, e);
 		}
